+++ Frameworks/Adium Framework/Source/AIListBookmark.m
@@ -414,19 +414,19 @@
 	if (inChat.unviewedMentionCount) {
 		// We contain mentions; display both this and the content count.
 		if (inChat.unviewedMentionCount > 1) {
-			statusMessage = [NSString stringWithFormat:AILocalizedString(@"%d mentions, %d messages", "Status message for a bookmark (>1 mention, >1 messages)"),
-							 inChat.unviewedMentionCount, inChat.unviewedContentCount];
+			statusMessage = [NSString stringWithFormat:AILocalizedString(@"%lu mentions, %lu messages", "Status message for a bookmark (>1 mention, >1 messages)"),
+							 (unsigned long)inChat.unviewedMentionCount, (unsigned long)inChat.unviewedContentCount];
 		} else if (inChat.unviewedContentCount > 1) {
-			statusMessage = [NSString stringWithFormat:AILocalizedString(@"1 mention, %d messages", "Status message for a bookmark (1 mention, >1 messages)"),
-							 inChat.unviewedContentCount];
+			statusMessage = [NSString stringWithFormat:AILocalizedString(@"1 mention, %lu messages", "Status message for a bookmark (1 mention, >1 messages)"),
+							 (unsigned long)inChat.unviewedContentCount];
 		} else {
 			statusMessage = AILocalizedString(@"1 mention, 1 message", "Status message for a bookmark (1 mention, 1 message)");
 		}
 	} else if (inChat.unviewedContentCount) {
 		// We don't contain mentions; display the content count.
 		if (inChat.unviewedContentCount > 1) {
-			statusMessage = [NSString stringWithFormat:AILocalizedString(@"%d messages", "Status message for a bookmark (>1 messages)"),
-							 inChat.unviewedContentCount];
+			statusMessage = [NSString stringWithFormat:AILocalizedString(@"%lu messages", "Status message for a bookmark (>1 messages)"),
+							 (unsigned long)inChat.unviewedContentCount];
 		} else {
 			statusMessage = AILocalizedString(@"1 message", "Status message for a bookmark (1 message)");
 		}
